requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
streamlit>=1.29.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
psycopg2-binary>=2.9.0
